rename contracts util to nft
