+++ packages/dapp/src/utils/constants/addresses.js
@@ -7,8 +7,8 @@ export const addresses = (chainId) => {
       }
     case 1337:
       return {
-        signatureFund: '0x5FC8d32690cc91D4c39d9d3abcBD16989F875707',
-        signatureNFT: '0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9',
+        signatureFund: '0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6',
+        signatureNFT: '0xa513E6E4b8f2a923D98304ec87F64353C4D5C853',
       }
     default:
       return {
