+++ packages/dapp/src/utils/nft.js
@@ -2,17 +2,12 @@ import { addresses, abis } from './constants'
 import { ethers, Contract } from 'ethers';
 
 export const createSign = async ({value, token, provider, signer}) => {
-
   const { chainId } = await provider.getNetwork()
   const signatureFundContract = new Contract(
       addresses(chainId).signatureFund,
       abis.signatureFund,
       signer)
-  return signatureFundContract.createSign(token, {
-    value: ethers.utils.parseEther(value),
-    gasLimit: 200000
-  })
-
+  return signatureFundContract.createSign(token, { value: ethers.utils.parseEther(value) })
 }
 
 export const mintSelected = async ({
